+++ CHANGELOG.md
@@ -4,9 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## 2.1.0 - 2019-2-21
 ### Fixed
-- Fixed [Issue 10](https://github.com/delightedCrow/WallberryTheme/issues/10): Background fades not reaching to bottom of screen for other 3rd party MM2 modules.
+- Fixed [Issue 10](https://github.com/delightedCrow/WallberryTheme/issues/10), where background fades did not reach to the edges of the screen for other 3rd party MM2 modules.
 - Provided a more permanent fix for [Issue 8](https://github.com/delightedCrow/WallberryTheme/issues/8) by adding an option to clear Electron's cache on start up.
 - WallberryTheme, WB-clock, and WB-weather all properly suspend their activity when hidden now, and resume properly when woken back up again.